+++ android/src/main/kotlin/com/hoanglm/flutteryoutubeview/YoutubeFactory.kt
@@ -2,19 +2,20 @@ package com.hoanglm.flutteryoutubeview
 
 import android.content.Context
 import androidx.lifecycle.Lifecycle
-import io.flutter.plugin.common.PluginRegistry
+import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.StandardMessageCodec
 import io.flutter.plugin.platform.PlatformView
 import io.flutter.plugin.platform.PlatformViewFactory
-import java.util.concurrent.atomic.AtomicReference
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.flow.StateFlow
 
 class YoutubeFactory(
-    private val registrar: PluginRegistry.Registrar,
-    private val state: AtomicReference<Lifecycle.Event>
+        private val binaryMessenger: BinaryMessenger,
+        private val lifecycleChannel: StateFlow<Lifecycle.Event>
 ) :
     PlatformViewFactory(StandardMessageCodec.INSTANCE) {
 
     override fun create(context: Context?, viewId: Int, args: Any?): PlatformView {
-        return FlutterYoutubeView(context!!, viewId, args as HashMap<String, *>, state, registrar)
+        return FlutterYoutubeView(context!!, viewId, args as HashMap<String, *>, binaryMessenger, lifecycleChannel)
     }
 }